feat(client): expose subscription event details
The Client.Update() method have been able to subscribe to pulseaudio
server state changes and send out notifications, but it ignores the
event type, facility and index information.

This adds the extraction of the event information from the packet, and
send them out via a channel. Also, the new Subscribe() method allows
specifying the event type to receive. The old Update() is changed to use
the Subscribe() method, while preserving the output.